DistanceCalculator to use Longs, distance is now a Long too.

+++ server/src/main/java/com/tco/requests/RequestDistance.java
@@ -7,7 +7,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class RequestDistance extends RequestHeader{
-    private Integer distance;
+    private Long distance;
     Float earthRadius;
     Map<String,String> place1;
     Map<String,String> place2;
@@ -30,7 +30,7 @@ public class RequestDistance extends RequestHeader{
         log.trace("buildResponse -> {}", this);
     }
 
-    public Integer getDistance() {
+    public Long getDistance() {
         return distance;
     }
